# src/public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyOrder API",
        "description": "MyOrder API",
        "contact": {
            "name": "ABhinav Gupta",
            "email": "abhinav.gupta02@nagarro.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/orders": {
            "post": {
                "tags": [
                    "Create New Order"
                ],
                "summary": "Place an order",
                "description": "Create a new Order after valid request of Geo-coordinates",
                "operationId": "placeOrder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "origin": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "destination": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "origin": [
                                        "28.644800",
                                        "77.308601"
                                    ],
                                    "destination": [
                                        "19.076090",
                                        "72.877426"
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Invalid Parameters"
                    }
                }
            }
        },
        "/orders/{orderID}": {
            "patch": {
                "tags": [
                    "Take a Order"
                ],
                "summary": "Take a order",
                "description": "Take order by Order Id where Order status is UNASSIGNED",
                "operationId": "takeOrder",
                "parameters": [
                    {
                        "name": "orderID",
                        "in": "path",
                        "description": "Valid Order Id",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "status": "TAKEN"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order successfully Taken"
                    },
                    "400": {
                        "description": "BAD_REQUEST"
                    },
                    "409": {
                        "description": "CONFLICT_ORDER_ALREADY_TAKEN"
                    },
                    "422": {
                        "description": "INVALID_PARAMETERS"
                    },
                    "405": {
                        "description": "METHOD_NOT_ALLOWED"
                    }
                }
            }
        },
        "/orders?page=:page&limit=:limit": {
            "get": {
                "tags": [
                    "Order List"
                ],
                "summary": "Return Order List",
                "description": "Returns Order List data have order id, status, disatnce",
                "operationId": "getOrderList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Valid Page Number",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Valid Order Limit",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "BAD_REQUEST"
                    },
                    "422": {
                        "description": "INVALID_PARAMETERS"
                    },
                    "405": {
                        "description": "METHOD_NOT_ALLOWED"
                    }
                }
            }
        }
    }
}